Be compatible with both old and versions of libwacom
libwacom 1.10 introduced a breaking change to the output format of
libwacom-list-local-devices. Commit a480df1 makes us compatible with the
new format, at the expense of breaking compatibility with older versions.

This commit reintroduces compatibility with BOTH versions. The strategy is
to first try running libwacom-list-local-devices with the new syntax, and
fall back to the old syntax if necessary.